[FRONT] Add page managerDelivery
